Get transition back to main state working
There were two problems: 1, I started the states counting at 0, but I
defined 0 as no state change, so it was ignoring it. 2, I wasn't having
the game state tick act as a subroutine call, so when the main menu
state subroutine did a rts, it went all the way back.